Client - Move EnumChip.vue and ActiveStatusChip.vue to the widgets folders. They were in the wrong place anyway since EntityTable was not the one making use of them, but they were rather passed to EntityTable from the respective Entity that called them up as a slot for EntityTable that is created viaa the TableHeaders -> Cellformatters
